+++ CHANGELOG.md
@@ -83,6 +83,7 @@
   - `file_name`
   - `first_where`
   - `flatmap_over_map_reduce`
+  - `for_where`
   - `force_try`
   - `force_unwrapping`
   - `function_body_length`
@@ -214,6 +215,12 @@
 
 * Print violations in realtime if `--progress` and `--output` are both set.  
   [JP Simard](https://github.com/jpsim)
+
+* Update `for_where` rule, adding a new configuration
+  `allow_for_as_filter` to allow using `for in` with a single `if` inside
+  when there's a `return` statement inside the `if`'s body.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#4040](https://github.com/realm/SwiftLint/issues/4040)
 
 * `quick_discouraged_call`, `quick_discouraged_focused_test` and
   `quick_discouraged_pending_test` rules now trigger on subclasses of